# docs/contracts/knowledge-v0.openapi.yaml
openapi: 3.0.3
info:
  title: TenderHack knowledge service — contract v0
  version: "0.1.0"
  description: >
    Frozen contract between the .NET support core (`api`/`api-worker`) and the Python
    knowledge & inference service (`knowledge`/`knowledge-worker`). See
    docs/contracts/knowledge-v0.md for normative rules (headers, error taxonomy,
    idempotency, versioning). This file is the source of truth for NSwag client
    generation until `knowledge` emits its own OpenAPI from FastAPI; at that point the
    two must stay in lockstep or the change is a contract-breaking change (see §6 of
    the companion markdown).
servers:
  - url: http://knowledge:8000
paths:
...
  /v0/understand:
    post:
      operationId: understand
      summary: Normalize raw user text and extract entities/exact codes.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - $ref: "#/components/parameters/XCaseId"
        - $ref: "#/components/parameters/XTurnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnderstandRequest"
      responses:
        "200":
          description: Understanding result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnderstandResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ModelUnavailable"
...
components:
  parameters:
    XTraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    XCaseId:
      name: X-Case-Id
      in: header
      required: true
      schema:
        type: string
    XTurnId:
      name: X-Turn-Id
      in: header
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Missing required trace header.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request body failed schema validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ModelUnavailable:
      description: Inference runtime unreachable — maps to KnowledgeFailure.UNAVAILABLE.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Unhandled error — maps to KnowledgeFailure.UNAVAILABLE.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - MISSING_TRACE_HEADER
            - UNKNOWN_SNAPSHOT
            - UNKNOWN_FRAGMENT
            - UNKNOWN_DOCUMENT
            - MODEL_UNAVAILABLE
            - MODEL_ERROR
            - INTERNAL_ERROR
        message:
          type: string
        detail:
          type: object
          additionalProperties: true
          nullable: true
...
    EntityProvenance:
      type: string
      enum: [user_explicit, trusted_portal_context, inferred, unknown]
    Entity:
      type: object
      required: [type, value, provenance]
      properties:
        type:
          type: string
          description: >
            Context slot name, e.g. role, process, edo_provider, document_type, status,
            user_action, error_code, duration, already_tried.
        value:
          type: string
        provenance:
          $ref: "#/components/schemas/EntityProvenance"
    LanguageFlags:
      type: object
      required: [detected_language, typo_corrected]
      properties:
        detected_language:
          type: string
        typo_corrected:
          type: boolean
    UnderstandRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        prior_turn_summary:
          type: string
          nullable: true
    UnderstandResponse:
      type: object
      required: [normalized_text, entities, exact_codes, language_flags]
      properties:
        normalized_text:
          type: string
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        exact_codes:
          type: array
          items:
            type: string
        language_flags:
          $ref: "#/components/schemas/LanguageFlags"
...